Use BigInteger for money and upgrade costs
Switch numeric types to System.Numerics.BigInteger for currency and upgrade costs to support very large values: Global.Money, _damageCost, _cooldownCost and _radiusCost now use BigInteger and relevant files import System.Numerics. The enemy reward calculation in HealthHandler was changed from a floating-point exponential to BigInteger.Pow (uses integer division Global.Difficulty/5), which will alter scaling (rewards now step every 5 difficulty). The upgrade cost function now returns BigInteger. Additionally, export_presets.cfg was updated (export_path adjusted and patch-delta related fields added/modified).

+++ Shop/Upgrades.cs
@@ -1,5 +1,6 @@
 using Godot;
 using System;
+using System.Numerics;
 
 namespace tdstopdownshooter.Shop;
 
@@ -10,13 +11,13 @@ public partial class Upgrades : VBoxContainer
     [Export] public Button RadiusButton;
     [Export] public Label MoneyLabel;
 
-    private long _damageCost = 5;
-    private long _cooldownCost = 5;
-    private long _radiusCost = 5;
+    private BigInteger _damageCost = 5;
+    private BigInteger _cooldownCost = 5;
+    private BigInteger _radiusCost = 5;
 
-    private static int Function(int timesBought)
+    private static BigInteger Function(int timesBought)
     {
-        return (int)Math.Round(5.0 * Math.Pow(5.0, 0.22 * timesBought), MidpointRounding.AwayFromZero);
+        return (BigInteger)Math.Round(5.0 * Math.Pow(5.0, 0.22 * timesBought), MidpointRounding.AwayFromZero);
     }
 
     private void UpdateData()
